# ipcache: Consolidate prefixes from various resources