--- a/pkg/policy/cell/policy_importer.go
+++ b/pkg/policy/cell/policy_importer.go
@@ -170,6 +170,7 @@ func (i *policyImporter) updatePrefixes(ctx context.Context, updates []*policyty
 					Source:   prefixSource[resource],
 					Resource: resource,
 					Metadata: []ipcache.IPMetadata{labels.GetCIDRLabels(prefix)},
+					IsCIDR:   true,
 				})
 			}
 			continue
@@ -202,6 +203,7 @@ func (i *policyImporter) updatePrefixes(ctx context.Context, updates []*policyty
 				Source:   prefixSource[resource],
 				Resource: resource,
 				Metadata: []ipcache.IPMetadata{labels.GetCIDRLabels(prefix)},
+				IsCIDR:   true,
 			})
 		}
 		if oldSet.Len() > 0 {
@@ -245,6 +247,7 @@ func (i *policyImporter) prunePrefixes(prunePrefixes map[ipcachetypes.ResourceID
 				Prefix:   cmtypes.NewLocalPrefixCluster(oldPrefix),
 				Resource: resource,
 				Metadata: []ipcache.IPMetadata{labels.Labels{}},
+				IsCIDR:   true,
 			})
 		}
 	}