Depends on https://github.com/cilium/cilium/pull/38851.

- **testutils/ipcache: Remove nonexistent methods**
  AllocateCIDRs() and ReleaseCIDRIdentitiesByCIDR() have both been removed
  from the codebase already, so remove the leftovers from this mock
  ipcache.

- **ipcache: Extend benchmark to test CIDR churn**
  Benchmark removing and re-adding CIDR prefixes as well using the batch
  metadata API, to simulate the usage from the policy importer.

- **ipcache: Consolidate prefixes from various resources**
  When many CIDRs are inserted into the ipcache via the metadata API (i.e.
  UpsertMetadataBatch() / UpsertMetadata()) and many of those CIDRs are
  present across many various resources such as CNPs (aka "duplicated"
  CIDRs across many policies), the ipcache ends up performing many
  redundant label merging operations via ToLabels() inside
  resolveIdentity(). See BenchmarkManyCIDREntries() for more context.
  
  To optimize the ipcache metadata structure and avoid storing duplicate
  instances of metadata (particularly CIDR prefixes from policies), we
  handle CIDR prefixes outside the cluster as a special case. Instead of
  tracking each CIDR by its originating policy, we consolidate them by
  treating each CIDR as originating from its own dedicated resource. This
  approach reduces duplication.
  
  The above is done by a refcount of each CIDR prefix that is destined for
  outside of the cluster.
  
  Benchmark results:
  
  ```
  $ go test -v ./pkg/ipcache -test.run '^$' -test.bench 'BenchmarkManyCIDREntries' -count 10 > old
  $ go test -v ./pkg/ipcache -test.run '^$' -test.bench 'BenchmarkManyCIDREntries' -count 10 > new
  $ benchstat old new
  name                old time/op    new time/op    delta
  ManyCIDREntries-16    35.4ms ± 4%     0.2ms ± 1%  -99.33%  (p=0.000 n=10+10)
  
  name                old alloc/op   new alloc/op   delta
  ManyCIDREntries-16    4.94MB ± 1%    0.22MB ± 0%  -95.52%  (p=0.000 n=10+10)
  
  name                old allocs/op  new allocs/op  delta
  ManyCIDREntries-16     33.4k ± 1%      4.0k ± 0%  -88.11%  (p=0.000 n=10+10)
  ```
  
  Suggested-by: Casey Callendrello <cdc@isovalent.com>
  Suggested-by: Sebastian Wicki <sebastian@isovalent.com>